refactor: 에디터 관련 컴포넌트 features와 widgets으로 분리
