+++ apps/frontend/src/features/editor/hooks/useEditorTitle.ts
