+++ apps/frontend/src/features/editor/index.ts
@@ -1,0 +1,3 @@
+export { Editor } from "@/features/editor/ui/Editor";
+export { EditorTitle } from "@/features/editor/ui/EditorTitle";
+export { EditorActionPanel } from "@/features/editor/ui/EditorActionPanel";
