+++ apps/frontend/src/features/editor/ui/Editor/prosemirror.css
