+++ apps/frontend/src/features/editor/ui/Editor/selectors/node-selector.tsx
@@ -13,8 +13,11 @@ import {
 import { EditorBubbleItem, useEditor } from "novel";
 import { Popover } from "@radix-ui/react-popover";
 
-import { Button } from "@/components/editor/ui/button";
-import { PopoverContent, PopoverTrigger } from "@/components/editor/ui/popover";
+import { Button } from "@/features/editor/ui/Editor/ui/button";
+import {
+  PopoverContent,
+  PopoverTrigger,
+} from "@/features/editor/ui/Editor/ui/popover";
 
 const items: SelectorItem[] = [
   {
@@ -100,7 +103,7 @@ export const NodeSelector = ({ open, onOpenChange }: NodeSelectorProps) => {
     <Popover modal={true} open={open} onOpenChange={onOpenChange}>
       <PopoverTrigger
         asChild
-        className="hover:bg-accent gap-2 rounded-none border-none focus:ring-0"
+        className="gap-2 rounded-none border-none hover:bg-accent focus:ring-0"
       >
         <Button size="sm" variant="ghost" className="gap-2">
           <span className="whitespace-nowrap text-sm">{activeItem.name}</span>
@@ -115,7 +118,7 @@ export const NodeSelector = ({ open, onOpenChange }: NodeSelectorProps) => {
               item.command(editor);
               onOpenChange(false);
             }}
-            className="hover:bg-accent flex cursor-pointer items-center justify-between rounded-sm px-2 py-1 text-sm"
+            className="flex cursor-pointer items-center justify-between rounded-sm px-2 py-1 text-sm hover:bg-accent"
           >
             <div className="flex items-center space-x-2">
               <div className="rounded-sm border p-1">
