+++ apps/frontend/src/features/editor/ui/Editor/slash-commands.tsx
