+++ apps/frontend/src/features/editor/ui/Editor/ui/icons/crazy-spinner.tsx
