+++ apps/frontend/src/features/editor/ui/Editor/ui/icons/font-serif.tsx
