+++ apps/frontend/src/features/editor/ui/Editor/ui/icons/index.tsx
