+++ apps/frontend/src/features/editor/ui/Editor/ui/icons/loading-circle.tsx
