+++ apps/frontend/src/features/editor/ui/Editor/ui/scroll-area.tsx
