+++ apps/frontend/src/widgets/EditorView/index.ts
@@ -1,0 +1,1 @@
+export { EditorView } from "@/widgets/EditorView/ui/EditorView";
